Implement 802.1ah I-tag for PBB mac-in-mac encapsulation
